feat: add queue integration with back-end api
