+++ .env.example
@@ -1,4 +1,4 @@
-API_URL=https://api.spacecoders.app/api
+API_URL=http://localhost:3333
 
 # Redis
 REDIS_HOST=
