+++ .gitignore
@@ -3,3 +3,4 @@ dist
 tmp
 
 .env
+.env.prod
