+++ src/config/redis.ts
@@ -1,4 +1,9 @@
-export default {
-  host: process.env.REDIS_HOST,
-  port: Number(process.env.REDIS_PORT) || 6379,
-}
+const REDIS_HOST = process.env.REDIS_HOST;
+const REDIS_PORT = Number(process.env.REDIS_PORT) || 6379;
+const REDIS_URL = `redis://${REDIS_HOST}:${REDIS_PORT}`;
+
+export {
+  REDIS_HOST,
+  REDIS_PORT,
+  REDIS_URL
+};
