+++ src/dtos/IDifficultyDTO.ts
@@ -1,0 +1,5 @@
+export enum IDifficultyDTO {
+  EASY = 'easy',
+  NORMAL = 'normal',
+  HARD = 'hard',
+}
