--- src/jobs/index.ts
@@ -1,1 +1,0 @@
-export { TestsCorrection } from './TestsCorrection';
