--- src/queue.ts
@@ -1,5 +1,0 @@
-import 'dotenv/config';
-
-import { Queue } from './lib/Queue';
-
-Queue.process();
